refactor(taxonomy-move): check if all ids are from taxonomies and
create utils file.
